32-bit signed integer is used everywhere for the OID value (which conforms to the binary protocol b/w Cache and CLR)

+++ isc.onec.adapter/V8Service.cs
@@ -212,7 +212,7 @@ namespace isc.onec.bridge {
 		/// <returns></returns>
 		private Response Unmarshal(object value) {
 			if (value is MarshalByRefObject) {
-				long oid = this.repository.Add(value);
+				int oid = this.repository.Add(value);
 				return new Response(ResponseType.OBJECT, oid);
 			} else if (value != null && value.GetType() == typeof(bool)) {
 				return new Response(((bool) value));
